refactor(frontend): make type stubs more organized
Will aid long-term maintenance as more types need to be added
