+++ frontend/types/stemmaweb/service/base/index.ts
@@ -1,0 +1,6 @@
+/** Wrapper object to return API responses with a clean error-handling. */
+export interface BaseResponse<T> {
+  success: boolean;
+  data?: T;
+  message?: string;
+}
