+++ frontend/types/stemmaweb/service/index.ts
@@ -1,0 +1,2 @@
+export * from './base';
+export * from './stemmarest';
